+++ src/test/generateAPIClientSpec.ts
@@ -23,7 +23,11 @@ test.before(async () => {
     }
 });
 test.beforeEach(async () => {
-    await fs.unlink(OUTPUT_FILE);
+    try {
+        await fs.unlink(OUTPUT_FILE);
+    } catch (error) {
+        // no op
+    }
 });
 
 test.serial("Should convert data successfully", async t => {